+++ src/MartinGeorgiev/Doctrine/ORM/Query/AST/Functions/ToNumber.php
@@ -7,7 +7,9 @@ namespace MartinGeorgiev\Doctrine\ORM\Query\AST\Functions;
 /**
  * Implementation of PostgreSQL to_number().
  *
- * @see https://www.postgresql.org/docs/17/functions-formatting.html
+ * Supports Roman numeral conversion via RN pattern (PostgreSQL 18+).
+ *
+ * @see https://www.postgresql.org/docs/18/functions-formatting.html
  * @since 3.3.0
  */
 class ToNumber extends BaseFunction