+++ tests/Unit/MartinGeorgiev/Doctrine/ORM/Query/AST/Functions/JsonStripNullsTest.php
@@ -19,14 +19,16 @@ class JsonStripNullsTest extends TestCase
     protected function getExpectedSqlStatements(): array
     {
         return [
-            'SELECT json_strip_nulls(c0_.jsonObject1) AS sclr_0 FROM ContainsJsons c0_',
+            'strips nulls with one parameter' => 'SELECT json_strip_nulls(c0_.jsonObject1) AS sclr_0 FROM ContainsJsons c0_',
+            'strips nulls with null_value_treatment parameter' => "SELECT json_strip_nulls(c0_.jsonObject1, 'true') AS sclr_0 FROM ContainsJsons c0_",
         ];
     }
 
     protected function getDqlStatements(): array
     {
         return [
-            \sprintf('SELECT JSON_STRIP_NULLS(e.jsonObject1) FROM %s e', ContainsJsons::class),
+            'strips nulls with one parameter' => \sprintf('SELECT JSON_STRIP_NULLS(e.jsonObject1) FROM %s e', ContainsJsons::class),
+            'strips nulls with null_value_treatment parameter' => \sprintf("SELECT JSON_STRIP_NULLS(e.jsonObject1, 'true') FROM %s e", ContainsJsons::class),
         ];
     }
 }